+++ packages/payment-integration-api/src/payment/payment-initialize-options.ts
@@ -3,3 +3,10 @@ import { PaymentRequestOptions } from './payment-request-options';
 export interface PaymentInitializeOptions extends PaymentRequestOptions {
     [key: string]: unknown;
 }
+
+export interface InitializePaymentOptions {
+    authorizationToken?: string;
+    customerMessage?: string;
+    referenceId?: string;
+    useStoreCredit?: boolean;
+}